requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
djangorestframework==3.10.2
djangorestframework-jsonapi==2.8.0
djangorestframework-stubs==1.0.0
inflection==0.3.1
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
